refactor(main): change fetch comics logic handling
